# API-Go/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.Body:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  role.AssignFeaturesRequest:
    properties:
      features:
        items:
          $ref: '#/definitions/role.FeatureAssignment'
        type: array
    required:
    - features
    type: object
  role.AssignRoleRequest:
    properties:
      role_id:
        type: string
    required:
    - role_id
    type: object
  role.CreateRoleRequest:
    properties:
      description:
        type: string
      display_name:
        maxLength: 100
        type: string
      name:
        maxLength: 50
        type: string
    required:
    - display_name
    - name
    type: object
  role.CreateScopeOverrideRequest:
    properties:
      expires_at:
        type: string
      feature_id:
        type: string
      granted_by:
        type: string
      is_allowed:
        type: boolean
      reason:
        type: string
      scope_type:
        enum:
        - GLOBAL
        - TRAINING_CENTER
        - AREA
        - OWN_FICHAS
        - OWN_SCHEDULE
        - OWN_PROFILE
        - OWN_FICHA_AS_LEARNER
        type: string
      user_id:
        type: string
    required:
    - feature_id
    - granted_by
    - reason
    - scope_type
    - user_id
    type: object
  role.FeatureAssignment:
    properties:
      feature_id:
        type: string
      scope_type:
        enum:
        - GLOBAL
        - TRAINING_CENTER
        - AREA
        - OWN_FICHAS
        - OWN_SCHEDULE
        - OWN_PROFILE
        - OWN_FICHA_AS_LEARNER
        type: string
    required:
    - feature_id
    - scope_type
    type: object
  role.ListFeaturesResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/role.FeatureAssignment'
        type: array
    type: object
  role.ListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/role.RoleResponse'
        type: array
      page:
        type: integer
      total:
        type: integer
    type: object
  role.RoleResponse:
    properties:
      description:
        type: string
      display_name:
        type: string
      id:
        type: string
      is_system_role:
        type: boolean
      name:
        type: string
    type: object
  role.ScopeOverrideResponse:
    properties:
      expires_at:
        type: string
      feature_id:
        type: string
      granted_by:
        type: string
      id:
        type: string
      is_allowed:
        type: boolean
      reason:
        type: string
      scope_type:
        type: string
      user_id:
        type: string
    type: object
  role.UpdateRoleRequest:
    properties:
      description:
        type: string
      display_name:
        type: string
    type: object
  role.UserRoleResponse:
    properties:
      assigned_by:
        type: string
      display_name:
        type: string
      id:
        type: string
      role_id:
        type: string
      role_name:
        type: string
      scope_type:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 'API REST de Identity and Access Management (IAM): autenticación JWT,
    RBAC, usuarios, roles, funcionalidades, sesiones y auditoría.'
  title: API JeussAirel (IAM)
  version: "1.0"
paths:
...
  /roles:
    get:
      description: Devuelve una lista paginada de roles.
      parameters:
      - description: Número de página (default 1)
        in: query
        name: page
        type: integer
      - description: Tamaño de página 1-100 (default 20)
        in: query
        name: page_size
        type: integer
      responses:
        "200":
          description: Roles obtenidos
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/role.ListResponse'
              type: object
        "403":
          description: Sin permisos (IDENTITY_ROLE_VIEW:read)
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Listar roles
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Registra un nuevo rol en el sistema.
      parameters:
      - description: Datos del rol
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/role.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Rol creado
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/role.RoleResponse'
              type: object
        "400":
          description: Datos inválidos
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Sin permisos (IDENTITY_ROLE_MANAGE:write)
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Crear rol
      tags:
      - Roles
  /roles/{id}:
    delete:
      parameters:
      - description: UUID del rol
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Rol eliminado
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Sin permisos (IDENTITY_ROLE_MANAGE:write)
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Rol no encontrado
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Eliminar rol
      tags:
      - Roles
    get:
      consumes:
      - application/json
      parameters:
      - description: UUID del rol
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Rol encontrado
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/role.RoleResponse'
              type: object
        "403":
          description: Sin permisos (IDENTITY_ROLE_VIEW:read)
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Rol no encontrado
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Obtener rol por ID
      tags:
      - Roles
    put:
      parameters:
      - description: UUID del rol
        in: path
        name: id
        required: true
        type: string
      - description: Campos a actualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/role.UpdateRoleRequest'
      responses:
        "200":
          description: Rol actualizado
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/role.RoleResponse'
              type: object
        "403":
          description: Sin permisos (IDENTITY_ROLE_MANAGE:write)
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Rol no encontrado
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Actualizar rol
      tags:
      - Roles
  /roles/{id}/features:
    get:
      parameters:
      - description: UUID del rol
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Funcionalidades del rol
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/role.ListFeaturesResponse'
              type: object
        "403":
          description: Sin permisos (IDENTITY_ROLE_VIEW:read)
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Rol no encontrado
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Listar funcionalidades de un rol
      tags:
      - Roles
    put:
      parameters:
      - description: UUID del rol
        in: path
        name: id
        required: true
        type: string
      - description: Lista de funcionalidades con ámbito
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/role.AssignFeaturesRequest'
      responses:
        "200":
          description: Funcionalidades asignadas
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Sin permisos (IDENTITY_ROLE_MANAGE:write)
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Rol no encontrado
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Asignar funcionalidades a un rol
      tags:
      - Roles
  /scope-overrides:
    post:
      parameters:
      - description: Detalle del override
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/role.CreateScopeOverrideRequest'
      responses:
        "201":
          description: Override creado
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/role.ScopeOverrideResponse'
              type: object
        "403":
          description: Sin permisos (IDENTITY_SCOPE_MANAGE:write)
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Crear override de scope
  /scope-overrides/{id}:
    delete:
      parameters:
      - description: UUID del override
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Override eliminado
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Sin permisos (IDENTITY_SCOPE_MANAGE:write)
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Override no encontrado
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Eliminar override de scope
      tags:
      - Roles
  /scope-overrides/users/{id}:
    get:
      parameters:
      - description: UUID del usuario
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Overrides obtenidos
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/role.ScopeOverrideResponse'
              type: object
        "403":
          description: Sin permisos (IDENTITY_SCOPE_MANAGE:write)
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Usuario no encontrado
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Listar overrides de scope de un usuario
      tags:
      - Roles
...
  /users/{id}/roles:
    get:
      parameters:
      - description: UUID del usuario
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Roles del usuario
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/role.UserRoleResponse'
              type: object
        "403":
          description: Sin permisos (IDENTITY_ROLE_VIEW:read)
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Usuario no encontrado
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Listar roles de un usuario
    post:
      parameters:
      - description: UUID del usuario
        in: path
        name: id
        required: true
        type: string
      - description: ID del rol
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/role.AssignRoleRequest'
      responses:
        "201":
          description: Rol asignado al usuario
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Sin permisos (IDENTITY_ROLE_ASSIGN:write)
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Usuario o rol no encontrado
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Asignar rol a un usuario
      tags:
      - Roles
  /users/{id}/roles/{userRoleId}:
    delete:
      parameters:
      - description: UUID del usuario
        in: path
        name: id
        required: true
        type: string
      - description: UUID de la asignación usuario-rol
        in: path
        name: userRoleId
        required: true
        type: string
      responses:
        "200":
          description: Rol removido del usuario
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Sin permisos (IDENTITY_ROLE_ASSIGN:write)
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Asignación no encontrada
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Quitar rol de un usuario
      tags:
      - Roles
securityDefinitions:
  BearerAuth:
    description: 'Token JWT de acceso. Uso: `Authorization: Bearer <access_token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"